fix: widen permute/pad kernel addressing to int64
The permute kernels index with `row_offsets * cols`, where `row_offsets` is
`program_id * BLOCK + tl.arange(...)` -- int32. Once `rows * cols` reaches 2**31
that product wraps and the access lands at a negative/aliased address.

The forward already widened its READ path (`src.to(tl.int64)`) but left the
store int32. The backward had no widening at all, on either its gather of
`grad` or its scatter into `output_buffer`. The scatter is the dangerous one:
it corrupts memory silently instead of faulting.

Reachable in practice. A DeepSeek-V3-16B EP=8 step with imbalanced routing put
1,141,504 rows x 2048 cols = 2,337,800,192 elements through these kernels, which
surfaced as NaN gradients across the whole model a dozen steps into training --
including in a control arm with no MXFP8 anywhere, since the dispatcher runs
regardless of expert precision.

Verified on MI350X (gfx950) with a reverted-fix control:

                    small shapes    rows*cols >= 2**31
  without the fix   fwd/bwd pass    fwd FAIL, bwd FAIL
  with the fix      fwd/bwd pass    fwd pass, bwd pass

Small shapes pass either way, which is why this survived until a run with ~65k
tokens per rank. End to end, the arm that previously went NaN at step 11 with
377/377 parameters non-finite now completes cleanly.

Adds a regression test at the boundary, skipped when the device lacks the ~12
GiB the shapes require.

+++ torchao/prototype/moe_training/ep/permute.py
@@ -297,8 +297,13 @@ def _triton_permute_fwd_kernel(
     )
 
     write_mask = (row_offsets[:, None] < out_rows) & (col_offsets[None, :] < cols)
+    # int64: `row_offsets` is int32 (program_id * BLOCK + arange), so
+    # `row_offsets * cols` wraps once out_rows*cols >= 2**31 and the store lands
+    # at a negative/aliased address. The READ path above was already widened via
+    # `.to(tl.int64)` on `src`; the write was not. At DSV3-16B under ragged
+    # routing this is reachable: 1,141,504 rows x 2048 cols = 2,337,800,192.
     tl.store(
-        out_ptr + row_offsets[:, None] * cols + col_offsets[None, :],
+        out_ptr + row_offsets[:, None].to(tl.int64) * cols + col_offsets[None, :],
         vals,
         mask=write_mask,
     )
@@ -399,15 +404,20 @@ def _triton_permute_bwd_kernel(
     row_offsets = row_pid * BLOCK_ROWS + tl.arange(0, BLOCK_ROWS)
     col_offsets = col_pid * BLOCK_COLS + tl.arange(0, BLOCK_COLS)
 
+    # int64 on every address term. The backward had NO widening at all -- neither
+    # the gather of `grad` nor the scatter into `output_buffer` -- so both wrap
+    # once rows*cols >= 2**31, and the scatter writes to aliased addresses, which
+    # is silent memory corruption rather than a fault. See the forward's store
+    # for the reachable magnitude.
     dest_rows = tl.load(
         permuted_indices_ptr + row_offsets,
         mask=row_offsets < grad_rows,
         other=PADDING_VALUE,
-    )
+    ).to(tl.int64)
 
     read_mask = (row_offsets[:, None] < grad_rows) & (col_offsets[None, :] < grad_cols)
     grad_values = tl.load(
-        grad_ptr + row_offsets[:, None] * grad_cols + col_offsets[None, :],
+        grad_ptr + row_offsets[:, None].to(tl.int64) * grad_cols + col_offsets[None, :],
         mask=read_mask,
         other=PADDING_VALUE,
     )
